# perf(ext/http): use ServeHandlerInfo class instead of object literal