@marcosc90 2023-08-10
@bartlomieju sure!

 `wrk`
```js
const addr = Deno.args[0] ?? "127.0.0.1:4500";
const [hostname, port] = addr.split(":");
const { serve } = Deno;

serve({ hostname, port: Number(port), reusePort: true }, (_req, info) => {
  return new Response(info.remoteAddr.transport);
});

```

class getter vs object literal

```js
class ServeHandlerInfo {
  #inner = null;
  constructor(inner) {
    this.#inner = inner;
  }
  get remoteAddr() {
    return this.#inner.remoteAddr;
  }
}

const inner = {
  get remoteAddr() {
    return { transport: "tcp" };
  },
};

Deno.bench("new ServeHandlerInfo", () => {
  const i = new ServeHandlerInfo(inner);
});

Deno.bench("{}", () => {
  const i = {
    get remoteAddr() {
      return inner.remoteAddr;
    }
  };
});
```